refactor: Split some functions
