+++ internal/arcgen/lang/go/generate_test.go
@@ -1,10 +1,7 @@
-//nolint:testpackage
 package arcgengo
 
 import (
-	"bytes"
 	"context"
-	goast "go/ast"
 	"io"
 	"os"
 	"testing"
@@ -14,13 +11,12 @@ import (
 
 	"github.com/kunitsucom/arcgen/internal/config"
 	"github.com/kunitsucom/arcgen/internal/contexts"
-	"github.com/kunitsucom/arcgen/pkg/errors"
 )
 
 //nolint:paralleltest
 func TestGenerate(t *testing.T) {
 	t.Run("success,tests", func(t *testing.T) {
-		ctx := contexts.WithArgs(context.Background(), []string{
+		ctx := contexts.WithOSArgs(context.Background(), []string{
 			"arcgen",
 			"--go-column-tag=dbtest",
 			"--go-method-name-table=GetTableName",
@@ -60,7 +56,7 @@ func TestGenerate(t *testing.T) {
 	})
 
 	t.Run("failure,no.errsource", func(t *testing.T) {
-		ctx := contexts.WithArgs(context.Background(), []string{
+		ctx := contexts.WithOSArgs(context.Background(), []string{
 			"arcgen",
 			"--go-column-tag=dbtest",
 			"--go-method-name-table=GetTableName",
@@ -82,7 +78,7 @@ func TestGenerate(t *testing.T) {
 	})
 
 	t.Run("failure,no.errsource", func(t *testing.T) {
-		ctx := contexts.WithArgs(context.Background(), []string{
+		ctx := contexts.WithOSArgs(context.Background(), []string{
 			"arcgen",
 			"--go-column-tag=dbtest",
 			"--go-method-name-table=GetTableName",
@@ -104,7 +100,7 @@ func TestGenerate(t *testing.T) {
 	})
 
 	t.Run("failure,no-such-file-or-directory", func(t *testing.T) {
-		ctx := contexts.WithArgs(context.Background(), []string{
+		ctx := contexts.WithOSArgs(context.Background(), []string{
 			"arcgen",
 			"--go-column-tag=dbtest",
 			"--go-method-name-table=GetTableName",
@@ -126,7 +122,7 @@ func TestGenerate(t *testing.T) {
 	})
 
 	t.Run("failure,directory.dir", func(t *testing.T) {
-		ctx := contexts.WithArgs(context.Background(), []string{
+		ctx := contexts.WithOSArgs(context.Background(), []string{
 			"arcgen",
 			"--go-column-tag=dbtest",
 			"--go-method-name-table=GetTableName",
@@ -145,100 +141,5 @@ func TestGenerate(t *testing.T) {
 		for _, src := range remainingArgs {
 			require.ErrorContains(t, Generate(ctx, src), "is a directory")
 		}
-	})
-}
-
-var _ buffer = (*testBuffer)(nil)
-
-type testBuffer struct {
-	WriteFunc  func(p []byte) (n int, err error)
-	StringFunc func() string
-}
-
-func (w *testBuffer) Write(p []byte) (n int, err error) {
-	return w.WriteFunc(p)
-}
-
-func (w *testBuffer) String() string {
-	return w.StringFunc()
-}
-
-//nolint:paralleltest
-func Test_generate(t *testing.T) {
-	t.Run("failure,os.OpenFile", func(t *testing.T) {
-		arcSrcSets := ARCSourceSetSlice{
-			&ARCSourceSet{
-				Filename: "tests/invalid-source-set",
-				ARCSourceSlice: []*ARCSource{
-					nil,
-				},
-			},
-		}
-		backup := fileExt
-		t.Cleanup(func() { fileExt = backup })
-		fileExt = ".invalid-source-set"
-		err := generate(arcSrcSets)
-		require.ErrorIs(t, err, errors.ErrInvalidSourceSet)
-	})
-}
-
-func newTestARCSourceSet() *ARCSourceSet {
-	return &ARCSourceSet{
-		PackageName: "testpkg",
-		ARCSourceSlice: []*ARCSource{
-			{
-				TypeSpec: &goast.TypeSpec{
-					Name: &goast.Ident{
-						Name: "Test",
-					},
-				},
-				StructType: &goast.StructType{
-					Fields: &goast.FieldList{
-						List: []*goast.Field{
-							{
-								Names: []*goast.Ident{
-									{
-										Name: "ID",
-									},
-								},
-								Tag: &goast.BasicLit{
-									Value: "`dbtest:\"id\"`",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func Test_sprint(t *testing.T) {
-	t.Parallel()
-	t.Run("failure,buffer", func(t *testing.T) {
-		t.Parallel()
-		buf := &testBuffer{
-			WriteFunc: func(_ []byte) (n int, err error) {
-				return 0, io.ErrClosedPipe
-			},
-			StringFunc: func() string {
-				return ""
-			},
-		}
-		arcSrcSet := newTestARCSourceSet()
-		err := fprintColumns(nil, buf, arcSrcSet)
-		require.ErrorIs(t, err, io.ErrClosedPipe)
-	})
-
-	t.Run("failure,File", func(t *testing.T) {
-		t.Parallel()
-		f := &testBuffer{
-			WriteFunc: func(_ []byte) (n int, err error) {
-				return 0, io.ErrClosedPipe
-			},
-		}
-		arcSrcSet := newTestARCSourceSet()
-		err := fprintColumns(f, bytes.NewBuffer(nil), arcSrcSet)
-		require.ErrorIs(t, err, io.ErrClosedPipe)
 	})
 }
